HDFC: Business Analyst | Micro Experience Program
In this project, a performance dashboard for the usage of credit and debit cards by major banks in India has been created. It includes two pages, the first page shows a high level view with Month over Month changes for various KPIs like Total Spend, Transactions, Volume of cards, Spend per card and Spend per Transaction for both debit and credit cards. The second page illustrates the market share analysis for each bank and card-type by showing the trends for each KPI for the selected bank vs the rest of the industry. The data for this dashboard has been scraped from the official website of rbi using selenium in python.